Version final

+++ codigo/src/main/java/mx/uady/sicei/service/AlumnoService.java
@@ -1,23 +1,32 @@
 package mx.uady.sicei.service;
+
+import java.net.Socket;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
 import org.springframework.transaction.annotation.Transactional;
+
+import jdk.javadoc.internal.doclets.formats.html.SourceToHTMLConverter;
+
 // import jdk.nashorn.internal.ir.Request;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
+
 import mx.uady.sicei.exception.*;
 import mx.uady.sicei.model.Alumno;
 import mx.uady.sicei.model.Usuario;
 import mx.uady.sicei.model.Equipo;
 import mx.uady.sicei.model.Licenciatura;
+import mx.uady.sicei.model.Tutoria;
 import mx.uady.sicei.model.request.AlumnoRequest;
 import mx.uady.sicei.repository.AlumnoRepository;
 import mx.uady.sicei.repository.UsuarioRepository;
 import mx.uady.sicei.repository.EquipoRepository;
-//import mx.uady.sicei.repository.TutoriaRepository;
+import mx.uady.sicei.repository.TutoriaRepository;
 
 @Service
 public class AlumnoService {
@@ -26,14 +35,14 @@ public class AlumnoService {
     @Autowired
     private AlumnoRepository alumnoRepository;
     
-    //@Autowired
-    //private TutoriaRepository tutoriaRepository;
-    
     @Autowired
     private UsuarioRepository usuarioRepository;
 
     @Autowired
     private EquipoRepository equipoRepository;
+    
+    @Autowired
+    private TutoriaRepository tutoriaRepository;
     // POST, PUT, GET, GET by ID, DELETE para entidad Alumno
 
     @Transactional // Crear una transaccion
@@ -50,8 +59,6 @@ public class AlumnoService {
             
         Alumno alumno = new Alumno();
         alumno.setNombre(request.getNombre());
-        
-
         if (request.getEquipo()!=null && request.getEquipo()>0){
             Equipo equipo = equipoRepository.findById(request.getEquipo()).orElseThrow(()-> new BadRequestException("El valor del equipo no existe"));
             alumno.setEquipo(equipo);
@@ -106,7 +113,6 @@ public class AlumnoService {
         List<Alumno> alumnos = new LinkedList<>();
         alumnoRepository.findAll().iterator().forEachRemaining(alumnos::add); // SELECT(id, nombre)
         return alumnos;
-        // return alumnoRepository.findAll();
     }
     // @Transactional()
 
@@ -120,14 +126,19 @@ public class AlumnoService {
     }
     @Transactional()
     public void delete(Integer id) {
+        // Delete del alumno debe validar que no existan tutorias correspondientes. Debe eliminar el Usuario también 
         Optional<Alumno> alumno = alumnoRepository.findById(id);
         if (alumno.isPresent()){
-            //Verificar conecciones con tutorias y equipo 
+            //Verificar conecciones con tutorias  
+            System.out.println(alumno.get());
+            List<Tutoria> tutoria = tutoriaRepository.findByAlumno(alumno.get());
+            System.out.println(tutoria);
 
-            //
+            if (tutoria.size() > 0)
+                throw new BadRequestException("El alumno esta enlazado con una tutoria");
+
             alumnoRepository.deleteById(id);
-            usuarioRepository.deleteById(alumno.get().getUsuario().getId());                       
-             
+            usuarioRepository.deleteById(alumno.get().getUsuario().getId());                            
         }
      }
